+++ test/api/api.js
@@ -249,15 +249,17 @@ QUnit.test('component can be subclassed externally', function(assert) {
   const Component = videojs.getComponent('Component');
   const ControlBar = videojs.getComponent('ControlBar');
 
-  const player = new (videojs.extend(Component, {
-    reportUserActivity() {},
+  class TestComponent extends Component {
+    reportUserActivity() {}
     textTracks() {
       return {
         addEventListener: Function.prototype,
         removeEventListener: Function.prototype
       };
     }
-  }))({
+  }
+
+  const player = new TestComponent({
     id() {},
     reportUserActivity() {}
   });
@@ -275,14 +277,15 @@ function testHelperMakeTag() {
 
 QUnit.test('should extend Component', function(assert) {
   const Component = videojs.getComponent('Component');
-  const MyComponent = videojs.extend(Component, {
+
+  class MyComponent extends Component {
     constructor() {
       this.bar = true;
-    },
+    }
     foo() {
       return true;
     }
-  });
+  }
 
   const myComponent = new MyComponent();
 
@@ -291,7 +294,8 @@ QUnit.test('should extend Component', function(assert) {
   assert.ok(myComponent.bar, 'the constructor function is used');
   assert.ok(myComponent.foo(), 'instance methods are applied');
 
-  const NoMethods = videojs.extend(Component);
+  class NoMethods extends Component {}
+
   const noMethods = new NoMethods({});
 
   assert.ok(noMethods.on, 'should extend component with no methods or constructor');